moves xandr converters to new file
