+++ adapters/jwplayer/utils_test.go
@@ -40,58 +40,6 @@ func TestParseBidderParams(t *testing.T) {
 	assert.Empty(t, params)
 }
 
-func TestGetXandrImpExt(t *testing.T) {
-	appnexusExt := GetXandrImpExt("1234")
-	var appnexusImp xandrImpExt
-	json.Unmarshal(appnexusExt, &appnexusImp)
-	assert.Equal(t, 1234, appnexusImp.Appnexus.PlacementID)
-
-	var badAppnexusImp xandrImpExt
-	badAppnexusExt := GetXandrImpExt("-/")
-	json.Unmarshal(badAppnexusExt, &badAppnexusImp)
-	assert.Empty(t, badAppnexusImp)
-}
-
-func TestSetXandrVideoExt(t *testing.T) {
-	video := &openrtb2.Video{}
-	SetXandrVideoExt(video)
-	assert.Empty(t, video.Ext)
-
-	video.Placement = openrtb2.VideoPlacementTypeInArticle
-	SetXandrVideoExt(video)
-	assert.NotNil(t, video.Ext)
-	var ext xandrVideoExt
-	json.Unmarshal(video.Ext, &ext)
-	assert.Equal(t, Outstream, ext.Appnexus.Context)
-
-	video.Ext = nil
-	video.Placement = openrtb2.VideoPlacementTypeInStream
-	video.StartDelay = openrtb2.StartDelayGenericPostRoll.Ptr()
-	SetXandrVideoExt(video)
-	assert.NotNil(t, video.Ext)
-	json.Unmarshal(video.Ext, &ext)
-	assert.Equal(t, PostRoll, ext.Appnexus.Context)
-}
-
-func TestGetXandrContext(t *testing.T) {
-	video := openrtb2.Video{}
-	assert.Equal(t, Unknown, GetXandrContext(video))
-
-	video.Placement = openrtb2.VideoPlacementTypeInBanner
-	video.StartDelay = openrtb2.StartDelayPreRoll.Ptr()
-	assert.Equal(t, Outstream, GetXandrContext(video))
-
-	video.Placement = openrtb2.VideoPlacementTypeInStream
-	assert.Equal(t, PreRoll, GetXandrContext(video))
-}
-
-func TestGetXandrContextFromStartdelay(t *testing.T) {
-	assert.Equal(t, PreRoll, GetXandrContextFromStartdelay(openrtb2.StartDelayPreRoll))
-	assert.Equal(t, MidRoll, GetXandrContextFromStartdelay(openrtb2.StartDelayGenericMidRoll))
-	assert.Equal(t, MidRoll, GetXandrContextFromStartdelay(openrtb2.StartDelay(5)))
-	assert.Equal(t, PostRoll, GetXandrContextFromStartdelay(openrtb2.StartDelayGenericPostRoll))
-}
-
 func TestIsOutstream(t *testing.T) {
 	assert.False(t, IsOutstream(openrtb2.VideoPlacementTypeInStream))
 
@@ -273,16 +221,4 @@ func TestMakeOrtbSegments(t *testing.T) {
 	expectedSegments := []openrtb2.Segment{{Value: "1"}, {Value: "2"}, {Value: "3"}}
 	assert.Len(t, segments, len(expectedSegments))
 	assert.ElementsMatch(t, segments, expectedSegments)
-}
-
-func TestWriteToXandrKeywords(t *testing.T) {
-	keyword := "key=value"
-	var jwpsegs []string
-	WriteToXandrKeywords(&keyword, jwpsegs)
-	assert.Equal(t, "key=value", keyword)
-
-	jwpsegs = append(jwpsegs, "80808080")
-	WriteToXandrKeywords(&keyword, jwpsegs)
-	assert.Equal(t, "key=value,jwpseg=80808080", keyword)
-
 }
